Add files via upload
This my first project of data analysis, I did an analysis of three branches of an enterprise of super markets, that recolect sales data for three months and I analyzed things like sales and customoer. I used the dataset of https://www.kaggle.com/datasets/aungpyaeap/supermarket-sales.
In a few days I going to share the same analysis but with MySQL and Tableau.