Add files via upload
This is my first relatively larger project in this world of data analysis, it is about the analysis of the sales of a supermarket company in a quarter, where its three branches are analyzed, as well as the behavior of the clients and other factors. A few months ago I published this project, but I made some changes.